// libs/features/collection/billing/src/lib/shared/models/allocation-summary-details.ts
import { BilingualText } from '@gosi-ui/core';

/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export class AllocationSummaryDetails {
  type: BilingualText = new BilingualText();
  amount: number = undefined;
}